// src/messageBrokers/Interfaces/topics.ts
export enum BrokerTopics {
  PAYMENT_FAILED = 'PAYMENT_FAILED',
  ORDER_CONFIRMED = 'ORDER_CONFIRMED',
  ORDER_STORE_ACCEPTED = 'ORDER_STORE_ACCEPTED',
  DELIVERY_ASSIGN = 'DELIVERY_ASSIGN',
  ORDER_DELIVERY_ASSIGNED = 'ORDER_DELIVERY_ASSIGNED',
  ORDER_READY_FOR_PICKUP = 'ORDER_READY_FOR_PICKUP',
  ORDER_DELIVERY_ARRIVED_STORE = 'ORDER_DELIVERY_ARRIVED_STORE',
  ORDER_DELIVERY_PICKED = 'ORDER_DELIVERY_PICKED',
  ORDER_DELIVERY_ARRIVED_USER = 'ORDER_DELIVERY_ARRIVED_USER',
  ORDER_DELIVERED = 'ORDER_DELIVERED',
  SHOP_DETAILS_UPDATED_COMMUNICATION = 'SHOP_DETAILS_UPDATED_COMMUNICATION',
  USER_DETAILS_UPDATED_COMMUNICATION = 'USER_DETAILS_UPDATED_COMMUNICATION',
  DELIVERY_AGENT_DETAILS_UPDATED_COMMUNICATION = 'DELIVERY_AGENT_DETAILS_UPDATED_COMMUNICATION',
  SEND_USER_LOGIN_OTP = 'SEND_USER_LOGIN_OTP',
  SEND_LOGIN_OTP_SMS_CHANNEL = 'SEND_LOGIN_OTP_SMS_CHANNEL',
  SEND_TO_SMS_CHANNEL = 'SEND_TO_SMS_CHANNEL',
  SEND_TO_EMAIL_CHANNEL = 'SEND_TO_EMAIL_CHANNEL',
  SEND_TO_PUSH_CHANNEL = 'SEND_TO_PUSH_CHANNEL',
  SEND_TO_WHATSAPP_CHANNEL = 'SEND_TO_WHATSAPP_CHANNEL',
}

export enum PaymentEventDetails {
  PAYMENT_SUCCESS = 'PAYMENT_SUCCESS',
  PAYMENT_FAILED = 'PAYMENT_FAILED',
}

export enum OrderEventDetails {
  ORDER_CREATED = 'ORDER_CREATED',
}